# metal : slight speed-up for add and mul kernels