M2 Ultra

- `master`

| model                          |       size |     params | backend    | ngl |    threads | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------: | ---------- | ---------------: |
| LLaMA 7B mostly Q8_0           |   6.67 GiB |     6.74 B | Metal      |   1 |          4 | pp 512     |    629.07 ± 0.21 |
| LLaMA 7B mostly Q4_0           |   3.56 GiB |     6.74 B | Metal      |   1 |          4 | pp 512     |    629.60 ± 0.34 |
| LLaMA 7B mostly Q4_1           |   3.95 GiB |     6.74 B | Metal      |   1 |          4 | pp 512     |    631.94 ± 0.41 |
| LLaMA 7B mostly Q8_0           |   6.67 GiB |     6.74 B | Metal      |   1 |          4 | tg 128     |     62.05 ± 0.22 |
| LLaMA 7B mostly Q4_0           |   3.56 GiB |     6.74 B | Metal      |   1 |          4 | tg 128     |     88.54 ± 0.04 |
| LLaMA 7B mostly Q4_1           |   3.95 GiB |     6.74 B | Metal      |   1 |          4 | tg 128     |     83.63 ± 0.04 |

- PR

| model                          |       size |     params | backend    | ngl |    threads | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------: | ---------- | ---------------: |
| LLaMA 7B mostly Q8_0           |   6.67 GiB |     6.74 B | Metal      |   1 |          4 | pp 512     |    637.97 ± 0.38 |
| LLaMA 7B mostly Q4_0           |   3.56 GiB |     6.74 B | Metal      |   1 |          4 | pp 512     |    639.64 ± 1.51 |
| LLaMA 7B mostly Q4_1           |   3.95 GiB |     6.74 B | Metal      |   1 |          4 | pp 512     |    643.86 ± 0.07 |
| LLaMA 7B mostly Q8_0           |   6.67 GiB |     6.74 B | Metal      |   1 |          4 | tg 128     |     62.34 ± 0.02 |
| LLaMA 7B mostly Q4_0           |   3.56 GiB |     6.74 B | Metal      |   1 |          4 | tg 128     |     88.28 ± 0.06 |
| LLaMA 7B mostly Q4_1           |   3.95 GiB |     6.74 B | Metal      |   1 |          4 | tg 128     |     83.69 ± 0.08 |