@jhen0409 2023-09-01
M2 (10c GPU) (use 4 threads)

- `master`

| model                          |       size |     params | backend    | ngl | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------- | ---------------: |
| models 7B mostly Q4_0          |   3.56 GiB |     6.74 B | Metal      |   1 | pp 512     |     90.25 ± 0.23 |
| models 7B mostly Q4_1          |   3.95 GiB |     6.74 B | Metal      |   1 | pp 512     |     90.53 ± 0.01 |
| models 7B mostly Q8_0          |   6.67 GiB |     6.74 B | Metal      |   1 | pp 512     |     90.17 ± 0.01 |
| models 7B mostly Q4_0          |   3.56 GiB |     6.74 B | Metal      |   1 | tg 128     |     20.73 ± 0.17 |
| models 7B mostly Q4_1          |   3.95 GiB |     6.74 B | Metal      |   1 | tg 128     |     17.57 ± 1.12 |
| models 7B mostly Q8_0          |   6.67 GiB |     6.74 B | Metal      |   1 | tg 128     |     10.99 ± 0.34 |

- PR

| model                          |       size |     params | backend    | ngl | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------- | ---------------: |
| models 7B mostly Q4_0          |   3.56 GiB |     6.74 B | Metal      |   1 | pp 512     |     90.00 ± 0.65 |
| models 7B mostly Q4_1          |   3.95 GiB |     6.74 B | Metal      |   1 | pp 512     |     90.77 ± 0.70 |
| models 7B mostly Q8_0          |   6.67 GiB |     6.74 B | Metal      |   1 | pp 512     |     90.82 ± 0.02 |
| models 7B mostly Q4_0          |   3.56 GiB |     6.74 B | Metal      |   1 | tg 128     |     21.26 ± 0.02 |
| models 7B mostly Q4_1          |   3.95 GiB |     6.74 B | Metal      |   1 | tg 128     |     19.38 ± 0.02 |
| models 7B mostly Q8_0          |   6.67 GiB |     6.74 B | Metal      |   1 | tg 128     |     12.03 ± 0.04 |